Remove obsolete SquareOneUnfilteredPuzzle
The default Square-1 solver is fast now, so we should use it. Also, in contrast to 4x4 random turns, the optimal solver initializes (rather) quickly and even with the previous UnfilteredPuzzle it would have been invoked anyways. That class was literally just there to limit the generic BFS solver, and said BFS isn't used anymore now. So might aswell remove this and encourage people to use actual scrambles.