Change Square-1 yellow to dark gray
Changed Square-1 yellow to dark gray by adding a new color, this is the most common square-1 color scheme, and fixes the problem with yellow and white contrasting, this would make it easier to check scrambles, and also make it easier on our eyes. The community has already voiced support for this change. https://forum.worldcubeassociation.org/t/change-clock-and-square-1-color-scheme-on-tnoodle-scramble-pdfs-to-white-black/42647